feat(wip): restructure app navigation flow
- Split MainActivity into separate activities
- Add dedicated SplashScreen in MainActivity
- Create new MainMenuActivity for menu navigation
- Implement initial app flow logic
- NOTE: Currently only working properly when started via debug mode
TODO: Fix app initialization in normal launch mode
