+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,9 @@
         android:theme="@style/Theme.Uncover"
         tools:targetApi="31">
         <activity
+            android:name=".MainMenuActivity"
+            android:exported="false" />
+        <activity
             android:name=".CharacterCreationActivity"
             android:exported="false" />
         <activity
